+++ Team5BLibraryManagementSystem/User.cs
@@ -14,7 +14,7 @@ namespace Team5BLibraryManagementSystem
     
     public partial class User
     {
-        public int userid { get; set; }
+        public double userid { get; set; }
         public string username { get; set; }
         public string password { get; set; }
     }
